fix: add user agent part with wakatime servers format
From an email of Alan Hamlett

You should change [the user agent] to: `helix/24.3 (2cadec0b) helix-wakatime/0.1.0`.

The last item in the User-Agent string is where we parse the plugin, and it currently has to follow the pattern `<editor>-wakatime/<plugin version>`. We have Helix as a known unreleased IDE, so it should show correctly on your dashboard with the above --plugin argument.